refactor(OBS-DELIVERY): split digest repo + clear arch-gate violations
Bring the branch under the ARCH-2 check:architecture gate (merged from main).

- Split observer-insight-repository.postgres.ts (938 lines): move the
  digest-delivery methods (claim/list/reserve/settle/recover + digest
  mappers) into a sibling observer-insight-repository.postgres.digest.ts as
  free functions taking the db handle; the main class delegates via thin
  forwarders (mirrors outbound-delivery-repository / canonical-message
  idiom). Shared schema refs + row mappers move to a
  observer-insight-repository.postgres.helpers.ts leaf both import. Pure
  relocation, behavior identical. Final: main 608, digest 422, helpers 54.
- observer-insight-emission.ts: reword a comment so the provider-boundary
  scanner no longer matches a provider token inside prose.
- runtime-services.ts: collapse one comment line in the observer-gateway
  wiring to return under its 1080-line budget.
- architecture-exceptions.json: add routes/observer.ts to the sanctioned
  forbidden_import_by_layer allowance peer control routes (settings, memory,
  jobs) already carry for resolving settings/derived status from config.
