+++ apps/core/src/adapters/storage/postgres/repositories/observer-insight-repository.postgres.ts
@@ -29,11 +29,26 @@ import type {
 import { isObserverSubjectKey } from '../../../../domain/ports/observer-insights.js';
 import * as pgSchema from '../schema/schema.js';
 import type { CanonicalDb } from './canonical-graph-repository.postgres.js';
+import {
+  Deliveries,
+  Insights,
+  clampLimit,
+  mapInsight,
+  nullableIso,
+  toIso,
+} from './observer-insight-repository.postgres.helpers.js';
+import {
+  claimPendingForDigest,
+  findDigestReservation,
+  findUnsettledDigestReservations,
+  listDigestDeliveries,
+  listPendingForDigest,
+  recoverStaleDigestClaims,
+  reserveDigest,
+  settleDigest,
+} from './observer-insight-repository.postgres.digest.js';
 
-const Insights = pgSchema.proactiveInsightsPostgres;
 const Embeddings = pgSchema.embeddingCachePostgres;
-const Deliveries = pgSchema.observerDeliveriesPostgres;
-const DeliveryInsights = pgSchema.observerDeliveryInsightsPostgres;
 const Cursors = pgSchema.observerInsightCursorsPostgres;
 const ACTIVE_INSIGHT_STATES: ObserverInsightState[] = [
   'pending',
@@ -389,155 +404,47 @@ export class PostgresObserverInsightRepository implements ObserverInsightReposit
     return mapDelivery(row);
   }
 
-  async claimPendingForDigest(input: {
+  claimPendingForDigest(input: {
     appId: string;
     recipient: string;
     limit: number;
     nowIso: string;
   }): Promise<ProactiveInsight[]> {
-    return this.db.transaction(async (tx) => {
-      const locked = await tx
-        .select({ id: Insights.id })
-        .from(Insights)
-        .where(
-          and(
-            eq(Insights.appId, input.appId),
-            eq(Insights.recipient, input.recipient),
-            eq(Insights.state, 'pending'),
-          ),
-        )
-        .orderBy(
-          desc(Insights.priorityScore),
-          asc(Insights.createdAt),
-          asc(Insights.id),
-        )
-        .limit(clampLimit(input.limit))
-        .for('update', { skipLocked: true });
-      if (locked.length === 0) return [];
-      const rows = await tx
-        .update(Insights)
-        .set({ state: 'claimed', updatedAt: input.nowIso })
-        .where(
-          inArray(
-            Insights.id,
-            locked.map((row) => row.id),
-          ),
-        )
-        .returning();
-      // UPDATE ... RETURNING has no defined order; membership positions depend
-      // on priority order, so re-sort by the ordered id sequence we locked.
-      const order = new Map(locked.map((row, index) => [row.id, index]));
-      return rows
-        .map(mapInsight)
-        .sort((left, right) => order.get(left.id)! - order.get(right.id)!);
-    });
+    return claimPendingForDigest(this.db, input);
   }
 
-  async listPendingForDigest(input: {
+  listPendingForDigest(input: {
     appId: string;
     recipient: string;
     limit: number;
   }): Promise<ProactiveInsight[]> {
-    const rows = await this.db
-      .select()
-      .from(Insights)
-      .where(
-        and(
-          eq(Insights.appId, input.appId),
-          eq(Insights.recipient, input.recipient),
-          eq(Insights.state, 'pending'),
-        ),
-      )
-      .orderBy(
-        desc(Insights.priorityScore),
-        asc(Insights.createdAt),
-        asc(Insights.id),
-      )
-      .limit(clampLimit(input.limit));
-    return rows.map(mapInsight);
+    return listPendingForDigest(this.db, input);
   }
 
-  async listDigestDeliveries(input: {
+  listDigestDeliveries(input: {
     appId: string;
     recipient: string;
     limit: number;
   }): Promise<ObserverDigestDeliverySummary[]> {
-    const insightCount = sql<number>`count(${DeliveryInsights.insightId})::int`;
-    const rows = await this.db
-      .select({
-        id: Deliveries.id,
-        localDay: Deliveries.localDay,
-        state: Deliveries.state,
-        reservedAt: Deliveries.reservedAt,
-        sentAt: Deliveries.sentAt,
-        settledAt: Deliveries.settledAt,
-        createdAt: Deliveries.createdAt,
-        insightCount,
-      })
-      .from(Deliveries)
-      .leftJoin(
-        DeliveryInsights,
-        eq(DeliveryInsights.deliveryId, Deliveries.id),
-      )
-      .where(
-        and(
-          eq(Deliveries.appId, input.appId),
-          eq(Deliveries.recipient, input.recipient),
-        ),
-      )
-      .groupBy(Deliveries.id)
-      .orderBy(desc(Deliveries.createdAt), desc(Deliveries.id))
-      .limit(clampLimit(input.limit));
-    return rows.map((row) => ({
-      id: row.id,
-      localDay: row.localDay,
-      state: row.state as ObserverDeliveryState,
-      insightCount: Number(row.insightCount ?? 0),
-      reservedAt: nullableIso(row.reservedAt),
-      sentAt: nullableIso(row.sentAt),
-      settledAt: nullableIso(row.settledAt),
-      createdAt: toIso(row.createdAt),
-    }));
+    return listDigestDeliveries(this.db, input);
   }
 
-  async findDigestReservation(input: {
+  findDigestReservation(input: {
     appId: string;
     recipient: string;
     localDay: string;
   }): Promise<ObserverDigestReservation | null> {
-    const [row] = await this.db
-      .select()
-      .from(Deliveries)
-      .where(
-        and(
-          eq(Deliveries.appId, input.appId),
-          eq(Deliveries.recipient, input.recipient),
-          eq(Deliveries.localDay, input.localDay),
-        ),
-      )
-      .limit(1);
-    return row ? mapReservation(row) : null;
+    return findDigestReservation(this.db, input);
   }
 
-  async findUnsettledDigestReservations(input: {
+  findUnsettledDigestReservations(input: {
     appId: string;
     recipient: string;
   }): Promise<ObserverDigestReservation[]> {
-    const rows = await this.db
-      .select()
-      .from(Deliveries)
-      .where(
-        and(
-          eq(Deliveries.appId, input.appId),
-          eq(Deliveries.recipient, input.recipient),
-          inArray(Deliveries.state, ['reserved', 'sent']),
-        ),
-      )
-      .orderBy(asc(Deliveries.createdAt), asc(Deliveries.id));
-    return rows.map(mapReservation);
+    return findUnsettledDigestReservations(this.db, input);
   }
 
-  async reserveDigest(input: {
+  reserveDigest(input: {
     id: string;
     appId: string;
     recipient: string;
@@ -551,198 +458,24 @@ export class PostgresObserverInsightRepository implements ObserverInsightReposit
     memberships: ObserverDigestClaimMembership[];
     nowIso: string;
   }): Promise<ObserverDigestReserveResult> {
-    return this.db.transaction(async (tx) => {
-      const [inserted] = await tx
-        .insert(Deliveries)
-        .values({
-          id: input.id,
-          appId: input.appId,
-          recipient: input.recipient,
-          localDay: input.localDay,
-          state: 'reserved',
-          timezone: input.timezone,
-          conversationJid: input.conversationJid,
-          providerAccountId: input.providerAccountId,
-          threadId: input.threadId ?? null,
-          renderedDigest: input.renderedDigest,
-          contentHash: input.contentHash,
-          reservedAt: input.nowIso,
-          createdAt: input.nowIso,
-        })
-        .onConflictDoNothing({
-          target: [Deliveries.appId, Deliveries.recipient, Deliveries.localDay],
-        })
-        .returning();
-
-      if (inserted) {
-        if (input.memberships.length > 0) {
-          await tx.insert(DeliveryInsights).values(
-            input.memberships.map((membership) => ({
-              deliveryId: inserted.id,
-              insightId: membership.insightId,
-              claimedAt: membership.claimedAt,
-              position: membership.position,
-            })),
-          );
-        }
-        return { reservation: mapReservation(inserted), created: true };
-      }
-
-      // Unique key already claimed this (app, recipient, day): the reservation
-      // is at-most-once, so return the existing one and add no membership.
-      const [existing] = await tx
-        .select()
-        .from(Deliveries)
-        .where(
-          and(
-            eq(Deliveries.appId, input.appId),
-            eq(Deliveries.recipient, input.recipient),
-            eq(Deliveries.localDay, input.localDay),
-          ),
-        )
-        .limit(1);
-      if (!existing) {
-        throw new Error('Observer digest reservation conflict without a row');
-      }
-      return { reservation: mapReservation(existing), created: false };
-    });
+    return reserveDigest(this.db, input);
   }
 
-  async settleDigest(input: {
+  settleDigest(input: {
     deliveryId: string;
     outboundDeliveryId: string;
     cooldownUntil: string;
     nowIso: string;
   }): Promise<ObserverDigestReservation | null> {
-    return this.db.transaction(async (tx) => {
-      const [delivery] = await tx
-        .select()
-        .from(Deliveries)
-        .where(eq(Deliveries.id, input.deliveryId))
-        .limit(1)
-        // Lock the delivery row so the state check below and the final
-        // transition are atomic. A concurrent settle blocks here, then reads
-        // the committed `settled` state and returns it idempotently instead of
-        // racing past the check and losing the compare-and-set (returning null).
-        .for('update');
-      if (!delivery) return null;
-      // Idempotent: an already-settled delivery is returned untouched (never
-      // overwrite its outboundDeliveryId/timestamps). Never settle a `failed`
-      // delivery.
-      if (delivery.state === 'settled') return mapReservation(delivery);
-      if (delivery.state !== 'reserved' && delivery.state !== 'sent') {
-        return null;
-      }
-
-      const memberships = await tx
-        .select()
-        .from(DeliveryInsights)
-        .where(eq(DeliveryInsights.deliveryId, input.deliveryId));
-
-      for (const membership of memberships) {
-        const [sent] = await tx
-          .update(Insights)
-          .set({
-            state: 'sent',
-            deliveryId: input.deliveryId,
-            surfacedAt: input.nowIso,
-            updatedAt: input.nowIso,
-          })
-          .where(
-            and(
-              eq(Insights.id, membership.insightId),
-              eq(Insights.state, 'claimed'),
-              eq(Insights.updatedAt, membership.claimedAt),
-            ),
-          )
-          .returning({ id: Insights.id });
-        if (!sent) {
-          // Fence didn't match: this member was NOT sent. Leaving it `claimed`
-          // under a settled delivery strands it forever (recover excludes
-          // settled-delivery members). Release it so recovery/next digest can
-          // reclaim it.
-          await tx
-            .update(Insights)
-            .set({ state: 'pending', updatedAt: input.nowIso })
-            .where(
-              and(
-                eq(Insights.id, membership.insightId),
-                eq(Insights.state, 'claimed'),
-              ),
-            );
-          continue;
-        }
-        // The row is locked by the update above for the rest of this txn, so
-        // the state='sent' guard alone is a sound fence for the cooldown step.
-        await tx
-          .update(Insights)
-          .set({
-            state: 'cooldown',
-            cooldownUntil: input.cooldownUntil,
-            updatedAt: input.nowIso,
-          })
-          .where(
-            and(
-              eq(Insights.id, membership.insightId),
-              eq(Insights.state, 'sent'),
-            ),
-          );
-      }
-
-      const [settled] = await tx
-        .update(Deliveries)
-        .set({
-          state: 'settled',
-          outboundDeliveryId: input.outboundDeliveryId,
-          sentAt: delivery.sentAt ?? input.nowIso,
-          settledAt: input.nowIso,
-        })
-        .where(
-          and(
-            eq(Deliveries.id, input.deliveryId),
-            inArray(Deliveries.state, ['reserved', 'sent']),
-          ),
-        )
-        .returning();
-      return settled ? mapReservation(settled) : null;
-    });
+    return settleDigest(this.db, input);
   }
 
-  async recoverStaleDigestClaims(input: {
+  recoverStaleDigestClaims(input: {
     appId: string;
     staleBeforeIso: string;
     nowIso: string;
   }): Promise<ProactiveInsight[]> {
-    const staleBefore = Date.parse(input.staleBeforeIso);
-    const recoveryTime = Date.parse(input.nowIso);
-    if (
-      !Number.isFinite(staleBefore) ||
-      !Number.isFinite(recoveryTime) ||
-      recoveryTime <= staleBefore
-    ) {
-      throw new Error(
-        'Observer claim recovery time must follow the stale cutoff',
-      );
-    }
-    const rows = await this.db
-      .update(Insights)
-      .set({ state: 'pending', updatedAt: input.nowIso })
-      .where(
-        and(
-          eq(Insights.appId, input.appId),
-          eq(Insights.state, 'claimed'),
-          lte(Insights.updatedAt, input.staleBeforeIso),
-          sql`${Insights.id} NOT IN (
-            SELECT ${DeliveryInsights.insightId}
-            FROM ${DeliveryInsights}
-            INNER JOIN ${Deliveries}
-              ON ${Deliveries.id} = ${DeliveryInsights.deliveryId}
-            WHERE ${Deliveries.state} IN ('reserved', 'sent', 'settled')
-          )`,
-        ),
-      )
-      .returning();
-    return rows.map(mapInsight);
+    return recoverStaleDigestClaims(this.db, input);
   }
 
   async getInsightCursor(
@@ -850,10 +583,6 @@ function keysetFilter(
   );
 }
 
-function clampLimit(limit: number): number {
-  return Math.max(1, Math.min(limit, 100));
-}
-
 function clampPageLimit(limit: number): number {
   return Math.max(1, Math.min(limit, 101));
 }
@@ -868,34 +597,6 @@ function assertCanonicalSubject(
   }
 }
 
-function mapInsight(row: typeof Insights.$inferSelect): ProactiveInsight {
-  return {
-    id: row.id,
-    appId: row.appId,
-    subject: row.subject as ObserverSubjectKey,
-    insightType: row.insightType as ProactiveInsight['insightType'],
-    title: row.title,
-    summary: row.summary,
-    evidenceRefs: Array.isArray(row.evidenceRefs)
-      ? (row.evidenceRefs as ProactiveInsight['evidenceRefs'])
-      : [],
-    batchSnapshotAt: toIso(row.batchSnapshotAt),
-    evidenceVersion: row.evidenceVersion,
-    canonicalSignature: row.canonicalSignature,
-    signatureEmbeddingRef: row.signatureEmbeddingRef ?? null,
-    confidence: row.confidence,
-    priorityScore: row.priorityScore,
-    state: row.state as ObserverInsightState,
-    cooldownUntil: nullableIso(row.cooldownUntil),
-    resolvedAt: nullableIso(row.resolvedAt),
-    surfacedAt: nullableIso(row.surfacedAt),
-    recipient: row.recipient,
-    deliveryId: row.deliveryId ?? null,
-    createdAt: toIso(row.createdAt),
-    updatedAt: toIso(row.updatedAt),
-  };
-}
-
 function mapDelivery(row: typeof Deliveries.$inferSelect): ObserverDelivery {
   return {
     id: row.id,
@@ -904,35 +605,4 @@ function mapDelivery(row: typeof Deliveries.$inferSelect): ObserverDelivery {
     localDay: row.localDay,
     createdAt: toIso(row.createdAt),
   };
-}
-
-function mapReservation(
-  row: typeof Deliveries.$inferSelect,
-): ObserverDigestReservation {
-  return {
-    id: row.id,
-    appId: row.appId,
-    recipient: row.recipient,
-    localDay: row.localDay,
-    state: row.state as ObserverDeliveryState,
-    timezone: row.timezone ?? null,
-    conversationJid: row.conversationJid ?? null,
-    providerAccountId: row.providerAccountId ?? null,
-    threadId: row.threadId ?? null,
-    renderedDigest: row.renderedDigest ?? null,
-    contentHash: row.contentHash ?? null,
-    outboundDeliveryId: row.outboundDeliveryId ?? null,
-    reservedAt: nullableIso(row.reservedAt),
-    sentAt: nullableIso(row.sentAt),
-    settledAt: nullableIso(row.settledAt),
-    createdAt: toIso(row.createdAt),
-  };
-}
-
-function nullableIso(value: string | null): string | null {
-  return value ? toIso(value) : null;
-}
-
-function toIso(value: string): string {
-  return new Date(value).toISOString();
 }
