+++ scripts/architecture-exceptions.json
@@ -504,6 +504,13 @@
     "maxViolations": 3
   },
   {
+    "file": "apps/core/src/control/server/routes/observer.ts",
+    "rule": "forbidden_import_by_layer",
+    "reason": "Baseline layer-boundary debt from the in-progress canonical runtime refactor; observer control route resolves owner/delivery status from config the same way peer routes (settings, memory, jobs) do; capped until the boundary slice removes this exception.",
+    "removeByPhase": "canonical-boundary-cleanup-phase",
+    "maxViolations": 2
+  },
+  {
     "file": "apps/core/src/domain/repositories/domain-types.ts",
     "rule": "forbidden_import_by_layer",
     "reason": "Baseline layer-boundary debt from the in-progress canonical runtime refactor; capped so new imports fail until the boundary slice removes this exception.",
